renamed dqn/util.py to utils.py and moved utils.py to genetic directory
